Complete rewrite of horoscope code to async
