+++ .env.example
@@ -1,5 +1,4 @@
 TOKEN=replace_me
 BING_API=replace_me
-DATAFILE=data.json
 LOGLEVEL=info
 #log levels: debug, info, warning, error, critical -- default: info
